Loading a JSON file now fills in the other fields.
You can load from one of two interfaces.
Hitting a load button will use one set of inputs.
And it will reformat that data and copy it to the other set of inputs.
Previously loading from the 3 fields _would_ copy to JSON.
But loading from JSON would _not_ copy to the other three fields.

Made some improvements to the error messages.
In paricular, when you enter a puzzle description and the program immediately sees a problem.
But the message are still only available in the JavaScript console.

Added several more samples.
Currently they are mostly in comments.
One is in a variable that isn't read yet.
You have to manually copy them into the text box,
then click "Load from JSON".